fastapi/backend/requirements.txt
fastapi
uvicorn[standard]
python-jose
